+++ changelog.md
@@ -2,6 +2,8 @@
 - add `IEnumerable<Option<T>>.ValuesIntoOrError`
 - add `IEnumerable<T>.TryFirst(predicate)`
 - add `ReadOnlySpan<T>.TryIndexOfAny` 
+- add `ErrorState(<E>).ToResult(Func<Result<T(, E)>>)`
+- add `ErrorState(<E>).ToResultAsync(Func<Task<Result<T(, E)>>>)`
 - make `TryIndexOf` and `TryLastIndexOf` generic 
 - mark obsolete `OptionsMarshall.TryGetValue` and `TryGetError` as errors (will be removed after 0.3.2)
 - updated TUnit to 1.17.11